avoid unsigned vs signed warning on win32 (build failure)
git-svn-id: https://xpra.org/svn/Xpra/trunk@12860 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/xpra/codecs/enc_x264/encoder.pyx
@@ -684,7 +684,7 @@ cdef class Encoder:
             self.file.flush()
         return cdata, client_options
 
-    def flush(self, int frame_no):
+    def flush(self, unsigned long frame_no):
         if self.frames>frame_no or self.context==NULL:
             return None, {}
         cdef int i = x264_encoder_delayed_frames(self.context)
